pyproject.toml
[project]
name = "ikaptcha"
version = "1.0.0"
description = "CRNN that solves the Ikariam pirate fortress captcha"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    # Inference (minimum to run scripts/predict_onnx.py)
    "onnxruntime>=1.20",
    "pillow>=10.0",
    "numpy>=1.26",
    # Training + PyTorch inference path
    "torch>=2.4",
    "torchvision>=0.19",
    # ONNX export
    "onnx>=1.17",
    "onnxscript>=0.1",
    # Real-data fetching
    "requests>=2.31",
    # YOLO baseline eval (cv2.dnn loads the original ONNX)
    "opencv-python>=4.8.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ikaptcha*"]

[tool.uv]
# Quick start:
#   uv venv
#   uv pip install -e .
